requirements.txt
# API and authentication
requests>=2.31.0
python-dotenv>=1.0.0
tenacity>=8.2.3
aiohttp

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Geospatial
geopy>=2.4.0
shapely>=2.0.2
folium>=0.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
click>=8.1.0
tqdm>=4.66.1
colorlog>=6.7.0

# Date/time handling
python-dateutil>=2.8.2

# Configuration
pyyaml>=6.0.1

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0